requirements.txt
fastapi == 0.115.7
supabase == 2.12.0
httpx == 0.28.1
uvicorn == 0.34.0
python-dotenv == 1.0.1
jinja2 == 3.1.5
python-multipart == 0.0.20